[settings] Fix erroneous controller teardown logs
The logic was wrong for when to report an error during teardown. When
the teardown stage is reached, the entry for that setting type should
already be removed from the mapping. The teardown phase was logging an
error if it didn't see it in the mapping, expecting it to still exist.

Fixed: 53458
Tests: fx -o setui_service_tests
Only effects logs, no new tests

Change-Id: I2c69b3929c6c9538198eb6db77cbe0f18d2152a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397601
Reviewed-by: Paul Faria <paulfaria@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>