[kernel][object] Ensure enumeration progress
This reworks the GetVmAspaceMaps and GetVmAspaceVmos to avoid
a race between performing a SoftFault and using the faulted in page.
Due to EnumerateChildren restarting from the beginning, for aspaces
with many mappings the faulted in page could get evicted before the
enumeration got back to the point of copying to it. As a result the
syscall could potentially spend many iterations making no progress.

By not considering the portion of the aspace that had already been
enumerated we can instead perform the fault+copy at the same time,
and then 'resume' the enumeration where we left of. Although the
enumeration may have to be re-walked several times, progress even
in the worst case is guaranteed.

Most of this commit ends up unifying the VmMapBuilder and
AspaceVmoEnumerator classes to share all the common logic instead
of continuing to duplicate it.

Changes to the ProcessGetInfoTest are simply undoing the error
code status changes that happend when the use of SoftFault was
introduced (If12c25f94c0d7e008f8fbdeefc007bc576d8082b)

Test: Correctness is covered by existing tests.

Bug: 53685

Change-Id: Ib500500e74e2f7fcc8c300ee0e267de74333fd04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397515
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>