[brcmfmac] Add sim support for ARP offload
Implement support for simulation of ARP offloading. Note that at
this time we don't actually generate ARP frames (as the firmware
does), but we do inhibit the relaying of frames to the driver when
appropriate. For now, this is sufficient for testing the cases we
care about.

Bug: 52305
Test: Included (arp_test.cc)
Change-Id: Ie3cdc00cf1baed217af4fc0793608eeb69d1f46b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396778
Testability-Review: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Joshua Conner <joshconner@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>