[sshutil] Use lower log levels
A lot of these logs show up pretty frequently and aren't actually a big
deal, especially for the OTA tests, which intentionally reboot the
device. So I'm lowering the log level for a bunch of these logs to make
them less frightening. If something actually unexpected happens, it
should be the calling code's responsibility to log the error at a higher
severity.

I also added an 's' to the end of the duration in the keepalive timeout
log to be clear that the duration is in seconds.

Bug: 48042
Change-Id: Ie38b8c96b1629ff44d71f942fe157e7a9a99fcdd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398259
Commit-Queue: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>