[brcmfmac] Don't enable ARP offload in non-promisc
In set_multicast_list, we were setting ARP offloading to the
inverse of promiscuous mode. This ends up overriding the default
interface setting, which can result in ARP transactions being
dropped (most notably, in AP mode).

To resolve this, we now only override the default setting when
promiscuous mode is enabled (which at the moment is never),
which is more consistent with the intent of the original logic.

As a side effect, replace the flags bitfield of the net_device
struct with a boolean since we were only using a single field
(IFF_UP).

Bug: 52305
Test: arp_test.cc
Change-Id: Id3d118f90e59fc7897bd374846abb6ffcc227709
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396780
Commit-Queue: Joshua Conner <joshconner@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Testability-Review: Zhiyi Chen <zhiyichen@google.com>