[modular][flake] Fix use-after-free in StopStoryCall
StoryProviderImpl, which owns `story_runtime_containers_`, may be
torn down before the async task in StopStoryCall that uses it
is scheduled.

The extra async task to destroy the StoryRuntimeController does not
seem to be necessary. None of the callbacks for StopStoryCall use
the StoryRuntimeController, and it's never called with the objects
that it's destroying in the stack (i.e. only called directly through
StoryProviderImpl).

Multiply: session_shell_test (fuchsia): 10
Test: fx test "fuchsia-pkg://fuchsia.com/modular_integration_tests#meta/session_shell_test.cmx" -- --gtest_filter=SessionShellTest.StoryControllerAnnotate --gtest_repeat=1000
Bug: 53597

Change-Id: Ia70b970812b0d1d6f2ed2d8005d461a62fbd3844
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396957
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Testability-Review: Gabe Schine <thatguy@google.com>