[bt][a2dp-sink] Respond to stale absolute volume requests
The Volume Relay for A2DP Sink stores the OnVolumeChanged responder and
only updates when the system volume changes. However, it is possible
that the peer that issues the OnVolumeChanged disconnects before
receiving the response. Consequently, the stored responder will continue
to exist in the Volume Relay. However, when a new peer connects to AVRCP
and A2DP and makes an OnVolumeChanged request, the FIDL stream gets
clogged, because the previous (stale) responder still hasn't been
resolved.

In practice, the hanging-get pattern requires only one outstanding
request at a time, so the onus is on the client to not further requests
until the first has been resolved.
However, this is not specified in the
[`fuchsia.bluetooth.avrcp.AbsoluteVolumeHandler`] documentation.
Regardless, it makes sense for the server-side to have an appropriate
policy, and it is logical and convenient to just report the current
volume to the previous responder if an additional OnVolumeChanged
request is received.

In production, this bug manifested when a remote device (i.e phone) was
streaming music, disconnected (while media was playing), forgot the
device, and attempted to repair + start streaming media. In this case,
media streaming was OK but volume control crashed for the reasons
mentioned above.

Bug: 53997
Test: Added test_volume_changes_with_peer_disconnection

Multiply: bt-a2dp-sink-tests

Change-Id: Iee719053cbde80bd538a1ff986b906c6a6ba0a1b