[isolated-ota] Emulate actual paver behavior
When ABR isn't supported, the paver service closes the boot manager
channel with an epitaph indicating it is not supported.  Due to
limitations in the go bindings, the go system updater had to interpret
multiple error conditionas as ABR not being supported, but the Rust
implementation is able to read/parse the epitaph.  This change modifies
the isolated ota tests to use the epitaph so it is compatible with the
Rust system updater.

Change-Id: I3d1fc80f2cec583703776c1815972700f503447c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398268
Reviewed-by: Sen Jiang <senj@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
Testability-Review: Simon Shields <simonshields@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>