[kernel][cmdline] Add parser support for kernel.ram.reserve
Extend the kernel command line parser to allow for a new argument
which allows users to request blocks of contiguous memory to be
reserved and the physical address published via the kernel command
line arguments to the rest of the system.

Reserved memory will be allocated and pinned, and then held for the
duration of the system's uptime.  It will not be available for the PMM
to share with anyone else.  Additionally, the region allocated may be used
to create VMOs via zx_vmo_create_physical.  It will not be subject to
the root resource filter (which lands in the CL after this one).

Also note that this is just the parser change.  The actual reservation
behavior comes in with the CL which introduces the root resource
filter.

This is all about supporting core testing.  See bug 34029 for details.

Test: new parser unit tests
Change-Id: I86c63f930cce7ad46af79771b28c7545edb8d2a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395566
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>