[codec_impl] Fetch Auxiliary Buffers
An auxiliary buffer collection is a set of secondary buffers that can be
used to provide additional context between two pipeline nodes. The
current primary use case for these buffers is for decryptors to copy the
clear input data through to the decoder. Typically, when the decryptor
is decrypting into secure buffers, the clear input data just gets copied
through to a secure buffer, so it is unable to be read from an
unprotected context. This data typically includes metadata information
that is useful to the decoder's parser that might be run in an
unprotected context.

Based on the CodecAdapter implementation, CodecImpl will now fetch
auxiliary buffers if necessary and provide mapped access to them via the
CodecBuffers that are shared with the CodecAdapters.

TEST: fx test codec_impl_unit_tests

Change-Id: Icf9c5f45be2c490924e05055075c7088f055a80e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396230
Commit-Queue: RJ Ascani <rjascani@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>