[tools/serial] use raw and non-lossy configuration
This change starts in a fix to the serial configuration, and descends
into removing the ring buffer usage, as that was dependent on the
VTIME configuration for forward progress.

The serial configuration now resembles cfmakeraw with hardware flow
control disabled and other assocoiated settings appropriate for our
use cases.

The removal of the VTIME stategy is because Go in general does not
expect Read() to return 0, nil. This discrepancy makes it both hard
to read the code, as well as hard to assert correct behavior.

The ring buffer usage is removed, as the use cases in which the unix
socket connection are used don't actually desire lossy buffering
behavior, and the locking strategy of the ring buffer is not a good
fit for goroutines hitting it with io.Copy on both sides. Instead we
always allocate an aux file even if the caller doesn't need one, and
we use that as our buffer to avoid keeping the whole buffer in
memory. Each unix socket client will receive a full copy of the file
and all subsequent updates until the server context is cancelled.

Bug: 46945
Change-Id: I475c2ca537adb07974977c8606efa5f01fc4a0f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385673
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>