[scripts][fuzzing] Rename CLI to Host
As the CLI class has grown, it has taken on more and more of the
platform-specific code, and the class formerly known as "Host" has
become more and more focused on the details of the build
environment.

This CL is the second part of a renaming to reflect that shift; it
renames "CommandLineInterface" to "Host".

Change-Id: Ib32d75cabf84a15b4c47f67ebff09bea8d39bb59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386241
Commit-Queue: Aaron Green <aarongreen@google.com>
Testability-Review: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>