[wlan][policy] Record connect failures
This change creates a function in SavedNetworksManager for recording
failures when attempting to connect to the network. The NetworkConfig
already contained a field for remembering failures to connect, but this
CL renames them to be more clear. There is no way to use the BSSID that
was in the existing field so it is removed for now.

Failures (and successes) are recorded in mod.rs which handles API calls
(without using the state machine currently) and inside of the client
state machine.

Change-Id: Id5c5faa768d9cf3a9a427f3a818ad3b883139b07
Bug: 53545
Test: added a unit test and modified existing tests:
	fx test wlancfg-tests
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397800
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Testability-Review: Kevin Sakuma <sakuma@google.com>