[kms-stateless] Add fallback for finding tee.
The old logic using fdio_watcher gets a ZX_ERR_NOT_SUPPORTED when run
inside of a component. I am assuming this has to do with the way devices
are presented to the sandbox.

This logic isn't critical so we just fallback to a fairly dumb polling
approach if we see that the watcher isn't supported.

Test: tested using a new component.

Change-Id: I5095fba215e953004b7dd8c45a057948fd9c17fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397793
Commit-Queue: Ed Coyne <edcoyne@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Testability-Review: Drew Fisher <zarvox@google.com>