[debugger] Improve ss
Inverts the search order of inline calls and disassembly. Previously
this would check the current line's address ranges disassemble that to
get physical function call addresses, and then look for inlines in
there. But this actually misses all inlines because the line address
lookup at the beginning by definition won't include them.

Now this searches for all inlines in the current block that are called
at the current line. Then it disassembles the data between those.

Previously the test did not call the full version and instead just
checked that the decode worked given a memory dump which is why the
problem was not noticed. Hooking up the full function required adding
additional mock data for line information and tests the new behavior.

Add a GetExtent() helper function for AddressRanges to compute the
union.

Add missing condition to get inline functions called within lexical
blocks.

Changes file/line comparison to not compare the compilation directory.
This was causing problems because the current LineTable implementation
omits the compilation directory and certain comparisons fail.

Tolerate zero line numbers coming from a DWARF file. The FileLine class
asserts that the line number is not 0. I found a file where the compiler
emitted a 0 line number for the call site of an inline function. This
doesn't have any meaning, but we need to tolerate whatever the compiler
generates. So this filters out 0 line numbers and empty file names when
creating FileLine information from the symbol file.

Change-Id: I7262c6169774c39141050f766a4cd0cc917d3938
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392453
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>