[kernel] Ensure RootJobObserver created after the root job has children.
Currently, the RootJobObserver --- which restarts the system if the root
job is ever without children --- is created before the root job actually
has any children.

We get away with this for the moment for two reasons:

  * We don't check for children in the RootJobObserver's OnInitialize
    callback;

  * The root job doesn't happen to set/clear any signals prior to its
    first child process being created, so the OnStateChange callback
    doesn't fire prior to the root job's children being created.

Upcoming changes to the observer API will prevent the
OnInitialize/OnStateChange checks being different.

This CL delays creation of the RootJobObserver until after the root
job's child process is created. We keep ownership of the RootJobObserver
with the Executor, but create a new function to allow userboot.cc to

Bug: 51463
Change-Id: I27b6eecb12429db0e348e3a69a79e1e63ad810ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395761
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>