[llcpp] Remove move constructor from OwnedSyncCallBase
This calls into fidl::Linearize, which I'm trying to remove.

It is also only used in tests and is expensive to call outside of tests.

yifeit@ mentioned that the move constructor is needed for C++14 support,
though we need further refactoring for that anyways.

Test: mostly removed tests, but fx test llcpp_types_tests
Change-Id: I21c2cb168df7e83ae02b5fb800f7ed606a989252
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398123
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>