[tools][netboot] Don't bind to debugPort when sending netboot query.
In order to discover a device, we bind to a netboot client port and send
a query to get the address corresponding to a specific nodename. One of
the ports in this range is the debug log port used by loglistener. This
occasionally causes loglistener to fail with "address already in use" if
another bot on that machine has already binded to that port.

Bug: 53262
Change-Id: Ie0db8c31f31294e2ff37e5e4f3c983c97bd3405a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398543
Commit-Queue: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>