[fdf] Fix logs in device recycle
During device destroy, fbl_recycle() method is called and within
this method, creation of Ref_Ptr of the device is not allowed.
See comments in RefCountedBase::AddRef() for reasoning. As this
is used in logs, the assert in AddRef() gets triggered. Changing
mkdevpath to fix this.

Test: fx test driver_host_tests
Change-Id: I9a25d962f25d0feeedca0404651d415d60e794dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398502
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>