[brcmfmac] Workaround for zero rssi signalling data.
The bug in firmware causes numerous zero rssi readings in
signalling data. This is due to the fact that firmware only
indicates the rssi in the first packet in an ampdu. The workaround
is to save the last non-zero value and uses it as long as the
signalled value is zero.

Test: Added a SIM test that tests driver's handling of 0 RSSI
values in signalling data.
Manually tested and do not see any zero indications in
the rssi stats (wlan iface stats).

Change-Id: Iba178739f4a0bf2b06f50b0f1b52fdd98a8db5f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390627
Commit-Queue: Karthik Krishnaswami <karthikrish@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>
Reviewed-by: Karl Ward <karlward@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Testability-Review: Kiet Tran <kiettran@google.com>