[scrutiny] Add command line argument parsing and forward request body.
Scrutiny in the long-term will need support for command line parsing.
This change introduces the ability to configure the port that scrutiny
runs on. It allows for the user to provide -p or --port option to set a
port. This change also forwards the request body to the dispatcher and
subsequent DataControllers.

Change-Id: I85e4e5b08e1365e28e368dec6dd5fae2fffbab1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397497
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Daniel Hwang <arkay@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>
Testability-Review: Benjamin Wright <benwright@google.com>