[audio_core] Wire device clock into the pipelines
With this CL, the device clock is passed up into the input
and output pipelines. A future CL will add reconciliation
between the client clock and the device clock at the place
where they meet, which will be the top-level mixer (closest
to the client). With this CL, the MixStage is directly
passed a reference clock for its destination stream, and all
other parties "ask their neighbor" for the clock -- a chain
that eventually leads to a device, MixStream or PacketQueue.

Fixed: 46629
Test: new OutputPipeline unittest

Change-Id: I7b2c376c630f1561405a083456029a2111eef00d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387459
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>