[zircon][inet6] Fix a buffer read overrun in ip6toa
This was located by ASAN during testing and ended up as flake-failing
an unrelated change that I was attempting to land.  ASAN provided a
pretty solid report, and the code itself allowed for a rather simple
repro.  Between the two of them, it was not too difficult to find the
actual mistake.

Restructure the code so that it is (hopefully) a bit easier to read
and understand, and a bit harder to make mistakes with.  Also add some
tests as there did not seem to be any before this.

Tests: new unit tests.
Change-Id: I83c1e2ba900b93e5127d5f4534001cfd5451f73a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399101
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Ed Coyne <edcoyne@google.com>
Testability-Review: Ed Coyne <edcoyne@google.com>