[sshutil] Make ConnectionError a struct
Previously it was a sentinel error:

  var ConnectionError = errors.New("SSH connection error")

However, a sentinel error can't wrap another arbitrary error, so it's
impossible to idiomatically determine the "cause" of a ConnectionError
using errors.Is or errors.As. This change makes it possible for
ConnectionError to wrap another error, which will be particularly useful
for making changes to the sshutil library to wrap errors in
ConnectionErrors without obscuring the underlying error that callers
sometimes check for.

Bug: 48042
Change-Id: If5cec9903e89ef082d2f527456c73eced7d072ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399624
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>