[tools][affected_builders] Add GN commands to tool
This adds a path to a GN binary as a parameter to affected_builders.

It also adds a function that takes the provided gn binary and runs `gn
analyze` with it.

The output of the analyze call is returned to the caller.

Change-Id: I112721e82b0930b0dff7a7469acf5adc66fed20d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399616
Testability-Review: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>