[hypervisor][arm64] Correctly restore active LRs
If an LR is not inactive, make sure that we correctly restore it when we
return to the guest.

This involves checking whether an interrupt in any state is in an LR.
This ensures that active interrupts are not added twice, and that
pending interrupts that have been popped from the tracker but not
serviced by the GIC are not lost.

Change-Id: I809ec0a3b90b8583ff6c12aeab0b6b5836a4b575
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399236
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>