[blobfs] Fix off-by-one error in Blob::SpaceAllocate
Blob::SpaceAllocate is reserving an extra node for each blob
in memory. This reduces the number of available nodes according
to the BlobFS process.

Note that this bug only affects reservations made against
the node map that is stored in-memory. It does not affect
the actual number of nodes that are allocated on disk. Hence
this bug does not appear if the BlobFS process is restarted,
since the in-memory state will be correctly initialized from disk.

Created new test /system/test/fs/blobfs-large-integration-max-reservation-test
which exercises this bug.

Bug: 54001
Test: Ran blobfs unit, integration and large integration tests
      (fx shell runtests -n blobfs-integration-test)
Change-Id: Icfc5e9ecd47db7c6e35bfdeb254b780152635a04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397999
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>