[kernel][thread] Have Thread own ThreadDispatcher
Thread should hold a RefPtr to ThreadDispatcher
instead of doing manual refcounting in
ThreadDispatcher. This requires some changes
to the exit path as Thread should be in charge
of the exit flow.

As a bonus, the DPC in ThreadDispatcher can be
folded into the DPC for Thread. It is still
necessary to create a DPC to cleanup the Thread
because a Thread cannot free itself while it's
running. We only need a single DPC to take
care of Thread and ThreadDispatcher together.

Fixed: 48779
TEST: runtests /system/test/core
      k ut all

Change-Id: I22da54c41c6d8a4377be0a55a20042a3f4408436
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374934
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>