[components] Only dispatch CapabilityRouted for namespace capabilities
Dispatching the ELF runner capability prior to the Started event was
really confusing and causing people to spend a lot more time debugging
tests than they needed to.

This CL ensures that the CapabilityRouted event is only dispatched
in BlackBoxTests for namespace capabilities.

In the future, if we wish to be able to mock or interpose runners,
we can introduce, a new, separate event type.

Change-Id: If1093afd0193200b6d8b83d0873c18a6380a3ea2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398965
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>