[rust] Add a Task abstraction to fuchsia-async
This is an initial change to get in place an API that we like. Later
it's likely that we'll reimplement these ideas and fuse them with the
implementation in Executor, but that's not needed just yet.

API is a reduced version of the smol::Task abstraction.

Two spawners are provided: spawn & local - for task spawning on the
current executor and on the current thread.

Tasks cancel work upon being dropped, unless they were 'detached' in
which case they continue executing in the background.

Tasks are themselves futures, and so can be await-ed on. This forms a
join() functionality.

Tasks are typed, and that type is returned from await.

MULTIPLY: fuchsia-async-tests: 100

Change-Id: Ie8c1209f147c07ef17ec5d3ab02295dcd0b8071e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397701
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Testability-Review: Adam Perry <adamperry@google.com>
Commit-Queue: Craig Tiller <ctiller@google.com>