[modular] Move some of per-agent init to AgentContextImpl constructor
In preparation to make session shells an agent, move init of all
services provided to agents, and launching of the agent component, to the constructor of AgentContextImpl. Session shells and regular
agents differ in what services are provided to them, and who has
responsibility for launching the sub-component, and this change
makes it easier to share the initialization code that they have in
common, and not the rest.

Bug: 53687
Change-Id: Ia12ef4fb9bef993de06e37739cfc6eb11a76c3b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/399306
Commit-Queue: Gabe Schine <thatguy@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>