[kernel] Add load_balancer class, load shed thresh
Create a load_balancer class. It is designed to be run periodically to
take a global look at the state of the system and make decisions to balance it.

The initial decision is to look at the load on all cpus and calculate a
threshold at which they should start load shedding.

Background: go/zir-loadshed
Test: Only unit test for now.

Change-Id: I2f3fc93f1a9c81c2bb909868c2143b0acddf323e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/362020
Commit-Queue: Ed Coyne <edcoyne@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>